Update and rename 2-4-cohens_d.md to 2-4-cohens_d.py
My solution for Cohen's D Statistic. This is mostly a Pandas selection. Also filtering out invalid births that were most likely not carried to term (less than 20 weeks). Is there a more standard way to do this with pandas instead of a list comprehension? Because I had to convert back to a Series and it did not seem efficient.